requirements.txt
# Server
fastapi
uvicorn[standard]
python-dotenv

# AI + RAG (PINNED for compatibility)
langchain==0.3.27
langchain-community==0.3.27
langchain-google-genai==2.1.10
langchain-anthropic==0.3.20
faiss-cpu==1.12.0

# Utilities
sentence-transformers
requests
pypdf
python-multipart